+++ app/src/main/java/eu/kanade/domain/DomainModule.kt
@@ -72,6 +72,7 @@ import tachiyomi.domain.entry.interactor.GetEntryWithChapters
 import tachiyomi.domain.entry.interactor.GetLibraryEntries
 import tachiyomi.domain.entry.interactor.GetMergedEntry
 import tachiyomi.domain.entry.interactor.GetNextUnreadChapter
+import tachiyomi.domain.entry.interactor.GetRelatedEntries
 import tachiyomi.domain.entry.interactor.NetworkToLocalEntry
 import tachiyomi.domain.entry.interactor.SetEntryCategories
 import tachiyomi.domain.entry.interactor.SetEntryChapterFlags
@@ -149,6 +150,7 @@ class DomainModule : InjektModule {
         addFactory { SetEntryCategories(get()) }
         addFactory { SetEntryChapterFlags(get()) }
         addFactory { NetworkToLocalEntry(get()) }
+        addFactory { GetRelatedEntries(get(), get()) }
         addFactory { EntryLibraryProgressResolver(entryLibraryProgressCalculators(get())) }
         addFactory { GetLibraryEntries(get(), get(), get(), get(), get(), get(), get()) }
         addFactory { SyncEntryWithSource(get(), get(), get(), get(), get(), get(), get()) }